+++ CHANGES.rst
@@ -5,6 +5,9 @@ v4.7.2
 
 Usability, bells and whistles
 -----------------------------
+* The Rebase editor is now aware of the ``drop``, ``break``, ``label``, ``merge`` and ``reset``
+  commands that were added in recent versions of Git.
+
 * Desktop notifications can now be enabled when pushing remotes by enabling the
   "Notify on Push" option in the preferences.
   (`#1404 <https://github.com/git-cola/git-cola/pull/1404>`_)